rename: featuers/game index 관리(#206)
