+++ src/features/game/components/calender/MatchCalendar.tsx
@@ -1,5 +1,5 @@
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui';
-import { CalenderBody, MatchCalendarCell } from '@/features/game//components';
+import { CalenderBody, MatchCalendarCell } from '@/features/game';
 import { useGetMatchScheduleQuery } from '@/features/game/apis/match-schedule/matchScheduleApi.query';
 import { useMatchStore } from '@/store/useMatchStore';
 import { format } from 'date-fns';
