--- src/features/game/components/index.ts
@@ -1,23 +1,0 @@
-/** components */
-// calender
-export { CalendarHeader } from '../components/calender/CalendarHeader';
-export { CalenderBody } from '../components/calender/CalenderBody';
-export { MatchCalendar } from '../components/calender/MatchCalendar';
-export { MatchCalendarCell } from '../components/calender/MatchCalendarCell';
-// card
-export { KeyRecordsCard } from '../components/card/KeyRecordsCard';
-//table
-export { BattingRecordTable } from '../components/table/BattingRecordTable';
-export { MatchScoreTable } from '../components/table/MatchScoreTable';
-export { MatchSummaryTable } from '../components/table/MatchSummaryTable';
-export { PitchingRecordTable } from '../components/table/PitchingRecordTable';
-// carousel
-export { CarouselCard } from '../components/carousel/CarouselCard';
-export { MatchInfoCarousel } from '../components/carousel/MatchInfoCarousel';
-// common
-export { MatchBoard } from '../components/common/MatchBoard';
-// watch-point
-export { StartingPitcherTable } from '../components/watch-point/StartingPitcherTable';
-export { TeamLineup } from '../components/watch-point/TeamLineup';
-export { TeamLineupAnimation } from '../components/watch-point/TeamLineupAnimation';
-export { StartingPitcherChart } from '../components/watch-point/StartingPitcherChart';
